requirements.txt
aiofiles==24.1.0
beautifulsoup4==4.12.3
crewai==0.41.1
crewai_tools==0.4.26
feedparser==6.0.11
gunicorn==22.0.0
html_sanitizer==2.4.4
langchain_openai==0.1.20
orjson==3.10.6
pydantic==2.8.2
python-dotenv==1.0.1
quart==0.19.6
Requests==2.32.3
selenium==4.23.1
utils==1.0.2
vaderSentiment==3.3.2
webdriver_manager==4.0.2
